4.0.0b2 with core Array API implementation
