+++ .readthedocs.yaml
@@ -5,17 +5,12 @@
 # Required
 version: 2
 
-# Set the OS, Python version and other tools you might need
 build:
   os: ubuntu-22.04
   tools:
     python: "3.12"
-    # You can also specify other tool versions:
-    # nodejs: "19"
-    # rust: "1.64"
-    # golang: "1.19"
 
-# Build documentation in the "docs/" directory with Sphinx
+
 sphinx:
   configuration: docs/source/conf.py
