+++ vectorgebra/__init__.py
@@ -2,10 +2,10 @@
     The numerical analysis library, fully implemented with Python.
 """
 
-from .linalg import *
+from . import linalg  # Hide complex methods under linalg name
 from .array import *
 from .math import *
 from .utils import *
-from .variable import Variable, grad, autograd, BinaryNode
+from .variable import Variable, grad, autograd
 
-__version__ = "4.0.0b1"
+__version__ = "4.0.0b2"
