+++ vectorgebra/linalg/__init__.py
@@ -1,0 +1,3 @@
+"""
+    This module contains more complex implementations for the Arrays
+"""
